documents: Rename files to follow the prefix-name rule
Renames the documents files to xdp-documents.
