+++ src/email.c
@@ -36,7 +36,7 @@
 
 #include "email.h"
 #include "xdp-request.h"
-#include "documents.h"
+#include "xdp-documents.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-utils.h"
