+++ src/meson.build
@@ -71,7 +71,6 @@ xdg_desktop_portal_sources = files(
   'background.c',
   'camera.c',
   'clipboard.c',
-  'documents.c',
   'dynamic-launcher.c',
   'email.c',
   'file-chooser.c',
@@ -100,6 +99,7 @@ xdg_desktop_portal_sources = files(
   'xdp-app-launch-context.c',
   'xdp-background-monitor.c',
   'xdp-call.c',
+  'xdp-documents.c',
   'xdp-permissions.c',
   'xdp-portal-impl.c',
   'xdp-request.c',
