+++ src/screenshot.c
@@ -37,7 +37,7 @@
 #include "screenshot.h"
 #include "xdp-permissions.h"
 #include "xdp-request.h"
-#include "documents.h"
+#include "xdp-documents.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-utils.h"
