+++ src/trash.c
@@ -36,7 +36,7 @@
 
 #include "trash.h"
 #include "xdp-call.h"
-#include "documents.h"
+#include "xdp-documents.h"
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
