+++ src/xdg-desktop-portal.c
@@ -31,7 +31,9 @@
 #include <glib/gi18n.h>
 
 #include "xdp-utils.h"
+#include "xdp-call.h"
 #include "xdp-dbus.h"
+#include "xdp-documents.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-method-info.h"
 #include "xdp-portal-impl.h"
@@ -39,10 +41,8 @@
 
 #include "account.h"
 #include "background.h"
-#include "xdp-call.h"
 #include "camera.h"
 #include "clipboard.h"
-#include "documents.h"
 #include "dynamic-launcher.h"
 #include "email.h"
 #include "file-chooser.h"
